+++ test/helpers/fake_data.h
@@ -1,0 +1,11 @@
+#pragma once
+
+#include "data/movie/Movie.h"
+
+#include <memory>
+
+namespace test {
+
+std::unique_ptr<Movie> movieWithAllDetails();
+
+} // namespace test